+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * new: Sequencex added maxOf\* and maxOf\* method
 * change: Mapx max\* and min\* method rename to max\*OrNull and min\*OrNull
 * change: Collectionx max\* and min\* method rename to max\*OrNull and min\*OrNull
+* change: Arrayx max\* and min\* method rename to max\*OrNull and min\*OrNull
 
 ## v1.0.0-beta02
 improve: Improve code